Player Movement Animation
Added Player Movement Animation

+++ Assessment_Assignment/include/gba_entity.h
@@ -54,7 +54,7 @@ typedef struct Player{
 	SpriteObject* sprite;
 	Vector2 screenPos; //Position on the screen
 	Vector2 worldPos; //Position in the world
-	u8 frame; //Current Animation Frame
+	u16 frame; //Current Animation Frame
 	fixed frameCounter; //Counter for current frame time
 
 	//Currently visible/pickup able
